requirements.txt
requests>=2.31.0
gspread>=6.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0
google-auth-httplib2>=0.2.0
